Moved EditableSection
