+++ amundsen_application/static/js/components/EditableSection/styles.scss
