+++ amundsen_application/static/js/pages/DashboardPage/index.tsx
@@ -16,7 +16,7 @@ import { updateSearchState } from 'ducks/search/reducer';
 
 import Breadcrumb from 'components/Breadcrumb';
 import BookmarkIcon from 'components/Bookmark/BookmarkIcon';
-import EditableSection from 'components/common/EditableSection';
+import EditableSection from 'components/EditableSection';
 import LoadingSpinner from 'components/common/LoadingSpinner';
 import TabsComponent, { TabInfo } from 'components/common/TabsComponent';
 import ResourceStatusMarker from 'components/common/ResourceStatusMarker';
