+++ amundsen_application/static/js/pages/TableDetailPage/index.tsx
@@ -34,7 +34,7 @@ import TabsComponent, { TabInfo } from 'components/common/TabsComponent';
 import TagInput from 'components/common/Tags/TagInput';
 import EditableText from 'components/common/EditableText';
 import LoadingSpinner from 'components/common/LoadingSpinner';
-import EditableSection from 'components/common/EditableSection';
+import EditableSection from 'components/EditableSection';
 import ColumnList from 'features/ColumnList';
 
 import { formatDateTimeShort } from 'utils/dateUtils';
